auth/requirements.txt
fastapi
uvicorn
passlib
bcrypt
python-jose
python-dotenv
redis_om
python-multipart